perf(plugin): reduce bundle size
