+++ .changeset/tame-ways-wink.md
@@ -1,0 +1,7 @@
+---
+'@modern-js/plugin': patch
+---
+
+perf(plugin): reduce bundle size
+
+perf(plugin): 优化包体积
