+++ packages/toolkit/plugin/src/farrow-pipeline/context.ts
@@ -2,9 +2,8 @@
  * modified from https://github.com/farrow-js/farrow/tree/master/packages/farrow-pipeline
  * license at https://github.com/farrow-js/farrow/blob/master/LICENSE
  */
+
 export type Context<T = any> = {
-  // create a new context equipped a new value
-  create: (value: T) => Context<T>;
   // get context ref { value } for accessing context in current container of pipeline
   use: () => {
     value: T;
@@ -15,32 +14,21 @@ export type Context<T = any> = {
   set: (value: T) => void;
 };
 
-export const createContext = <T>(value: T) => {
-  let currentValue: T;
+export const createContext = <T>(value: T): Context<T> => {
+  let currentValue = value;
 
-  const create = (value: T): Context<T> => {
-    currentValue = value;
-
-    const use = () => ({
+  return {
+    use: () => ({
       get value() {
         return currentValue;
       },
       set value(v) {
         currentValue = v;
       },
-    });
-    const get = () => currentValue;
-    const set = (v: T) => {
+    }),
+    get: () => currentValue,
+    set: (v: T) => {
       currentValue = v;
-    };
-    const Context: Context<T> = {
-      create,
-      use,
-      get,
-      set,
-    };
-    return Context;
+    },
   };
-
-  return create(value);
 };
